ai-agent/requirements.txt
web3==6.15.1
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
